Teh solo
Tea  branded with the name Solo City, Central Java, are starting to mushroom in various regions in Indonesia because tea from Bengawan City is believed to have its own uniqueness.
Talking about the city of Solo, of course we cannot escape its culinary delights and one of the things that is legendary to this day is the concocted tea drink which tastes of ginastel, an acronym for legi, hot and kenthel. In Indonesian it means sweet, hot and thick. Apart from ginastel, tea from Solo is also famous for its aroma and astringency.